+++ controls-magix/src/jvmTest/kotlin/space/kscience/controls/client/MagixLoopTest.kt
@@ -1,8 +1,13 @@
 package space.kscience.controls.client
 
-import kotlinx.coroutines.flow.*
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.flow.filterIsInstance
+import kotlinx.coroutines.flow.launchIn
+import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.test.runTest
+import kotlinx.coroutines.runBlocking
+import org.junit.jupiter.api.Timeout
 import space.kscience.controls.api.DescriptionMessage
 import space.kscience.controls.client.RemoteDeviceConnect.TestDevice
 import space.kscience.controls.manager.DeviceManager
@@ -16,14 +21,14 @@ import space.kscience.magix.rsocket.rSocketWithWebSockets
 import space.kscience.magix.server.startMagixServer
 import kotlin.test.Test
 import kotlin.test.assertEquals
-import kotlin.time.Duration.Companion.seconds
+import kotlin.time.Duration.Companion.milliseconds
 
 class MagixLoopTest {
 
     @Test
-    fun realDeviceHub() = runTest(timeout = 3.seconds) {
+    @Timeout(5)
+    fun realDeviceHub(): Unit = runBlocking {
         val context = Context {
-//            coroutineContext(Dispatchers.Default)
             plugin(DeviceManager)
         }
 
@@ -39,7 +44,7 @@ class MagixLoopTest {
             repeat(10) {
                 deviceManager.install("test[$it]", TestDevice)
             }
-        }.join()
+        }
 
         val clientEndpoint = MagixEndpoint.rSocketWithWebSockets("localhost")
 
@@ -47,22 +52,21 @@ class MagixLoopTest {
             .map { it.second }
             .filterIsInstance<DescriptionMessage>()
             .onEach { println(it) }
-            .launchIn(backgroundScope)
+            .launchIn(this)
+
 
         val remoteHub = clientEndpoint.remoteDeviceHub(context, "client", "device")
 
         assertEquals(0, remoteHub.devices.size)
         clientEndpoint.requestDeviceUpdate("client", "device")
 
-        // wait for the message with configuration
-        if(remoteHub.devices.isEmpty()) {
-            clientEndpoint.subscribe(DeviceManager.magixFormat, originFilter = listOf("device")).first()
-        }
 
-
+        delay(100.milliseconds)
 
         assertEquals(10, remoteHub.devices.size)
 
-        context.close()
+        clientEndpoint.close()
+        deviceEndpoint.close()
+        server.stop()
     }
 }