+++ controls-modbus/src/jvmMain/kotlin/space/kscience/controls/modbus/DeviceProcessImage.kt
@@ -12,8 +12,12 @@ import kotlinx.io.Buffer
 import space.kscience.controls.api.Device
 import space.kscience.controls.api.PropertyChangedMessage
 import space.kscience.controls.ports.readShort
-import space.kscience.controls.spec.*
+import space.kscience.controls.spec.DevicePropertySpec
+import space.kscience.controls.spec.useProperty
+import space.kscience.controls.spec.write
+import space.kscience.controls.spec.writeAsync
 import space.kscience.dataforge.io.Binary
+import kotlin.time.Duration.Companion.milliseconds
 import space.kscience.dataforge.meta.MetaConverter
 import space.kscience.dataforge.meta.MetaReader
 import kotlin.time.Duration.Companion.milliseconds
@@ -36,7 +40,7 @@ public class DeviceProcessImageBuilder<D : Device> internal constructor(
 
     public fun bind(
         key: ModbusRegistryKey.Coil,
-        propertySpec: MutableDevicePropertySpec<D, Boolean>,
+        propertySpec: DevicePropertySpec<Boolean>,
     ): ObservableDigitalOut = bind(key) { coil ->
         coil.addObserver { _, _ ->
             device.writeAsync(propertySpec, coil.isSet)
@@ -58,7 +62,7 @@ public class DeviceProcessImageBuilder<D : Device> internal constructor(
 
     public fun bind(
         key: ModbusRegistryKey.DiscreteInput,
-        propertySpec: DevicePropertySpec<D, Boolean>,
+        propertySpec: DevicePropertySpec<Boolean>,
     ): DigitalIn = bind(key) { input ->
         device.useProperty(propertySpec) { value ->
             input.set(value)
@@ -77,7 +81,7 @@ public class DeviceProcessImageBuilder<D : Device> internal constructor(
 
     public fun bind(
         key: ModbusRegistryKey.InputRegister,
-        propertySpec: DevicePropertySpec<D, Short>,
+        propertySpec: DevicePropertySpec<Short>,
     ): SimpleInputRegister = bind(key) { input ->
         device.useProperty(propertySpec) { value ->
             input.setValue(value)
@@ -96,7 +100,7 @@ public class DeviceProcessImageBuilder<D : Device> internal constructor(
 
     public fun bind(
         key: ModbusRegistryKey.HoldingRegister,
-        propertySpec: MutableDevicePropertySpec<D, Short>,
+        propertySpec: DevicePropertySpec<Short>,
     ): ObservableRegister = bind(key) { register ->
         register.addObserver { _, _ ->
             device.writeAsync(propertySpec, register.toShort())
@@ -134,7 +138,7 @@ public class DeviceProcessImageBuilder<D : Device> internal constructor(
     }
 
 
-    public fun <T> bind(key: ModbusRegistryKey.InputRange<T>, propertySpec: DevicePropertySpec<D, T>) {
+    public fun <T> bind(key: ModbusRegistryKey.InputRange<T>, propertySpec: DevicePropertySpec<T>) {
         val registers = List(key.count) {
             SimpleInputRegister()
         }