+++ demo/motors/src/main/kotlin/ru/mipt/npm/devices/pimotionmaster/PiMotionMasterApp.kt
@@ -19,11 +19,16 @@ import androidx.compose.ui.window.application
 import androidx.compose.ui.window.rememberWindowState
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.launch
+import space.kscience.controls.api.Device
+import space.kscience.controls.api.ParentDevice
 import space.kscience.controls.manager.DeviceManager
 import space.kscience.controls.manager.installing
+import space.kscience.controls.spec.execute
 import space.kscience.controls.spec.read
 import space.kscience.dataforge.context.Context
 import space.kscience.dataforge.context.request
+import space.kscience.dataforge.meta.Meta
+import space.kscience.dataforge.names.Name
 
 //class PiMotionMasterApp : App(PiMotionMasterView::class)
 //
@@ -42,25 +47,25 @@ import space.kscience.dataforge.context.request
 
 @Composable
 fun ColumnScope.piMotionMasterAxis(
-    axisName: String,
-    axis: PiMotionMasterDevice.Axis,
+    axisName: Name,
+    axis: Device,
 ) {
     var min by remember { mutableStateOf(0f) }
     var max by remember { mutableStateOf(1f) }
     var targetPosition by remember { mutableStateOf(0f) }
-    val position: Double by axis.composeState(PiMotionMasterDevice.Axis.position, 0.0)
+    val position: Double by axis.composeState(Axis.position, 0.0)
 
     val scope = rememberCoroutineScope()
 
     LaunchedEffect(axis) {
-        min = axis.read(PiMotionMasterDevice.Axis.minPosition).toFloat()
-        max = axis.read(PiMotionMasterDevice.Axis.maxPosition).toFloat()
-        targetPosition = axis.read(PiMotionMasterDevice.Axis.position).toFloat()
+        min = axis.read(Axis.minPosition).toFloat()
+        max = axis.read(Axis.maxPosition).toFloat()
+        targetPosition = axis.read(Axis.position).toFloat()
     }
 
 
     Row {
-        Text(axisName)
+        Text(axisName.toString())
 
         Column {
             Slider(
@@ -74,7 +79,7 @@ fun ColumnScope.piMotionMasterAxis(
                 onValueChange = { newPosition ->
                     targetPosition = newPosition
                     scope.launch {
-                        axis.move(newPosition.toDouble())
+                        axis.execute(Axis.move, Meta(newPosition))
                     }
                 },
                 valueRange = min..max
@@ -85,7 +90,7 @@ fun ColumnScope.piMotionMasterAxis(
 }
 
 @Composable
-fun AxisPane(axes: Map<String, PiMotionMasterDevice.Axis>) {
+fun AxisPane(axes: Map<Name, Device>) {
     Column {
         axes.forEach { (name, axis) ->
             this.piMotionMasterAxis(name, axis)
@@ -95,12 +100,12 @@ fun AxisPane(axes: Map<String, PiMotionMasterDevice.Axis>) {
 
 
 @Composable
-fun PiMotionMasterApp(device: PiMotionMasterDevice) {
+fun PiMotionMasterApp(device: ParentDevice) {
 
 //    val scope = rememberCoroutineScope()
-    val connected by device.composeState(PiMotionMasterDevice.connected, false)
+    val connected by device.composeState(PiMotionMaster.connected, false)
     var debugServerJob by remember { mutableStateOf<Job?>(null) }
-    var axes by remember { mutableStateOf<Map<String, PiMotionMasterDevice.Axis>?>(null) }
+    var axes by remember { mutableStateOf<Map<Name, Device>?>(null) }
     //private val axisList = FXCollections.observableArrayList<Map.Entry<String, PiMotionMasterDevice.Axis>>()
     var host by remember { mutableStateOf("127.0.0.1") }
     var port by remember { mutableStateOf(10024) }
@@ -159,12 +164,15 @@ fun PiMotionMasterApp(device: PiMotionMasterDevice) {
                     onClick = {
                         if (!connected) {
                             device.launch {
-                                device.connect(host, port)
-                                axes = device.axes
+                                device.execute(PiMotionMaster.connect, Meta {
+                                    "host" put host
+                                    "port" put port
+                                })
+                                axes = device.devices
                             }
                         } else {
                             device.launch {
-                                device.disconnect()
+                                device.execute(PiMotionMaster.disconnect)
                                 axes = null
                             }
                         }
@@ -197,7 +205,7 @@ fun main() = application {
     val deviceManager: DeviceManager = context.request(DeviceManager)
 
     // install device
-    val motionMaster: PiMotionMasterDevice by deviceManager.installing(PiMotionMasterDevice)
+    val motionMaster: ParentDevice by deviceManager.installing(PiMotionMaster)
 
     Window(
         title = "Pi motion master demo",