+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - New interface `WithLifeCycle`. Change Port API to adhere to it.
 
 ### Changed
+- **BREAKING** Full refactor of device spec and device builders. Now device spec is a separate content. There is a `DeviceFactory` that could be used to create both specification and device factory at the same time. 
 - Add optional names to model states and submodels
 - `StateContainer` renamed to `Constructor`
 - Separate StateContainer and MutableStateContainer